refactor(validation): Move logic for not-null field validation
Create implementation of the generic abstract class Validator to validate
that field is not null, named FieldNotNull.

FieldNotNull accept single or multiple fields upon construction.

If multiple fields are provided, it validates whether any of the
fields are null and returns an error message if so.

Delete specific not-null field validator classes.